Move test database code into its own module
Create ceilometer.tests.db to hold the TestConnection
and a test base class that sets one up.

Change the API test base class to inherit from the new
one that manages the DB handle.

Add a drop_database() method to TestConnection and implement
it using clear_all() for MIM databases to avoid recreating
the underlying Database() instance and its spidermonkey
interpreter instance, since we can only create a limited
number of spidermonkey instances in any one process.

Move the logic for deciding when to skip a test that needs
the Mongo map-reduce feature into the new db module.

Change-Id: Ic06f421b4caf472230bbb28637d6b7253808721e
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>